Refactor to be customizable.
