+++ src/ArrayExporter.php
@@ -1,0 +1,126 @@
+<?php
+/**
+ * Exporter
+ *
+ * Copyright (c) 2001-2014, Sebastian Bergmann <sebastian@phpunit.de>.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the name of Sebastian Bergmann nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @package    Exporter
+ * @author     Sebastian Bergmann <sebastian@phpunit.de>
+ * @copyright  2001-2014 Sebastian Bergmann <sebastian@phpunit.de>
+ * @license    http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
+ * @link       https://github.com/sebastianbergmann/exporter
+ */
+
+namespace SebastianBergmann\Exporter;
+
+/**
+ * Exporter for visualizing arrays.
+ *
+ * @package    Exporter
+ * @author     Sebastian Bergmann <sebastian@phpunit.de>
+ * @copyright  2001-2014 Sebastian Bergmann <sebastian@phpunit.de>
+ * @license    http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
+ * @link       https://github.com/sebastianbergmann/exporter
+ */
+class ArrayExporter extends BaseExporter
+{
+    /**
+     * Returns whether the exporter can export a given value.
+     *
+     * @param  mixed $value The value to export.
+     * @return boolean
+     */
+    public function accepts($value)
+    {
+        return is_array($value);
+    }
+
+    /**
+     * Recursively exports a value as a string.
+     *
+     * @param  mixed $value The value to export
+     * @param  integer $indentation The indentation level of the 2nd+ line
+     * @param  SebastianBergmann\Exporter\Context $processed Contains all objects and arrays that have previously been rendered
+     * @return string
+     * @see    SebastianBergmann\Exporter\Exporter::export
+     */
+    protected function recursiveExport(&$value, $indentation, $processed = NULL)
+    {
+        if (!$processed) {
+            $processed = new Context;
+        }
+
+        if (($key = $processed->contains($value)) !== FALSE) {
+            return 'Array &' . $key;
+        }
+
+        $key = $processed->add($value);
+        $whitespace = str_repeat(' ', 4 * $indentation);
+        $values = '';
+
+        if (count($value) > 0) {
+            foreach ($value as $k => $v) {
+                $keyExporter = $this->factory->getExporterFor($k);
+                $valueExporter = $this->factory->getExporterFor($value[$k]);
+
+                $values .= sprintf(
+                  '%s    %s => %s' . "\n",
+                  $whitespace,
+                  $keyExporter->recursiveExport($k, $indentation),
+                  $valueExporter->recursiveExport(
+                    $value[$k], $indentation + 1, $processed
+                  )
+                );
+            }
+
+            $values = "\n" . $values . $whitespace;
+        }
+
+        return sprintf('Array &%s (%s)', $key, $values);
+    }
+
+    /**
+     * Exports a value into a single-line string.
+     *
+     * @param  mixed $value
+     * @return string
+     * @see    SebastianBergmann\Exporter\Exporter::export
+     */
+    public function shortenedExport($value)
+    {
+        return sprintf(
+          'Array (%s)',
+          count($value) > 0 ? '...' : ''
+        );
+    }
+}
