+++ src/Factory.php
@@ -1,0 +1,102 @@
+<?php
+/**
+ * Exporter
+ *
+ * Copyright (c) 2001-2014, Sebastian Bergmann <sebastian@phpunit.de>.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the name of Sebastian Bergmann nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @package    Exporter
+ * @author     Sebastian Bergmann <sebastian@phpunit.de>
+ * @copyright  2001-2014 Sebastian Bergmann <sebastian@phpunit.de>
+ * @license    http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
+ * @link       https://github.com/sebastianbergmann/exporter
+ */
+
+namespace SebastianBergmann\Exporter;
+
+/**
+ * Factory of exporters for visualizing PHP variables.
+ *
+ * @package    Exporter
+ * @author     Sebastian Bergmann <sebastian@phpunit.de>
+ * @copyright  2001-2014 Sebastian Bergmann <sebastian@phpunit.de>
+ * @license    http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
+ * @link       https://github.com/sebastianbergmann/exporter
+ */
+class Factory
+{
+    /**
+     * @var BaseExporter[]
+     */
+    private $exporters = array();
+
+    /**
+     * Returns the correct exporter for exporting a given value.
+     *
+     * @param  mixed $value The value to export.
+     * @return BaseExporter
+     */
+    public function getExporterFor($value)
+    {
+        foreach ($this->exporters as $exporter) {
+            if ($exporter->accepts($value)) {
+                return $exporter;
+            }
+        }
+    }
+
+    /**
+     * Registers a new exporter.
+     *
+     * @param BaseExporter $exporter The registered exporter
+     */
+    public function register(BaseExporter $exporter)
+    {
+        array_unshift($this->exporters, $exporter);
+    }
+
+    /**
+     * Unregisters an exporter.
+     *
+     * This exporter will no longer be returned by getExporterFor().
+     *
+     * @param BaseExporter $exporter The unregistered exporter
+     */
+    public function unregister(BaseExporter $exporter)
+    {
+        foreach ($this->exporters as $key => $_exporter) {
+            if ($exporter === $_exporter) {
+                unset($this->exporters[$key]);
+            }
+        }
+    }
+}
