+++ src/SplObjectStorageExporter.php
@@ -1,0 +1,100 @@
+<?php
+/**
+ * Exporter
+ *
+ * Copyright (c) 2001-2014, Sebastian Bergmann <sebastian@phpunit.de>.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the name of Sebastian Bergmann nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @package    Exporter
+ * @author     Sebastian Bergmann <sebastian@phpunit.de>
+ * @copyright  2001-2014 Sebastian Bergmann <sebastian@phpunit.de>
+ * @license    http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
+ * @link       https://github.com/sebastianbergmann/exporter
+ */
+
+namespace SebastianBergmann\Exporter;
+
+/**
+ * Exporter for visualizing \SplObjectStorage instances.
+ *
+ * @package    Exporter
+ * @author     Sebastian Bergmann <sebastian@phpunit.de>
+ * @copyright  2001-2014 Sebastian Bergmann <sebastian@phpunit.de>
+ * @license    http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
+ * @link       https://github.com/sebastianbergmann/exporter
+ */
+class SplObjectStorageExporter extends ObjectExporter
+{
+    /**
+     * Returns whether the exporter can export a given value.
+     *
+     * @param  mixed $value The value to export.
+     * @return boolean
+     */
+    public function accepts($value)
+    {
+        return $value instanceof \SplObjectStorage;
+    }
+
+    /**
+     * Converts a PHP value to an array.
+     *
+     * @param  mixed $value
+     * @return array
+     */
+    public function toArray($value)
+    {
+        $array = parent::toArray($value);
+
+        // Remove HHVM internal representations
+        if (property_exists('\SplObjectStorage', '__storage')) {
+          unset($array['__storage']);
+        } else if (property_exists('\SplObjectStorage', 'storage')) {
+          unset($array['storage']);
+        }
+        if (property_exists('\SplObjectStorage', '__key')) {
+          unset($array['__key']);
+        }
+
+        // Some internal classes like SplObjectStorage don't work with our
+        // normal method of converting objects to arrays.
+        // Format the output similarly to print_r() in this case
+        foreach ($value as $key => $val) {
+            $array[spl_object_hash($val)] = array(
+                'obj' => $val,
+                'inf' => $value->getInfo(),
+            );
+        }
+
+        return $array;
+    }
+}
